Improved logic of helper files for package removal - files are not overwritten in updates / re-installation processes

+++ sbin/postinstallcheck.sh
@@ -1,20 +1,32 @@
 #!/bin/bash -e
 
-if [ ! -f /etc/init.d/newrelic-sysmond ];
+if [ ! -f /usr/local/psa/var/modules/new-relic/removepackageservers ]
 then
-    echo "1" > "/usr/local/psa/var/modules/new-relic/removepackageservers"
+    if [ ! -f /etc/init.d/newrelic-sysmond ];
+    then
+        echo "1" > "/usr/local/psa/var/modules/new-relic/removepackageservers"
+    else
+        echo "0" > "/usr/local/psa/var/modules/new-relic/removepackageservers"
+    fi
 fi
 
-if [ -f /etc/redhat-release ];
+if [ ! -f /usr/local/psa/var/modules/new-relic/removepackageapm ]
 then
-    if [ $(rpm -q newrelic-php5 2>/dev/null | grep -c "not installed") -eq 1 ];
+    if [ -f /etc/redhat-release ];
     then
-        echo "1" > "/usr/local/psa/var/modules/new-relic/removepackageapm"
-    fi
-else
-    if [ $(dpkg-query -W -f='${Status}' newrelic-php5 2>/dev/null | grep -c "ok installed") -eq 0 ];
-    then
-        echo "1" > "/usr/local/psa/var/modules/new-relic/removepackageapm"
+        if [ $(rpm -q newrelic-php5 2>/dev/null | grep -c "not installed") -eq 1 ];
+        then
+            echo "1" > "/usr/local/psa/var/modules/new-relic/removepackageapm"
+        else
+            echo "0" > "/usr/local/psa/var/modules/new-relic/removepackageapm"
+        fi
+    else
+        if [ $(dpkg-query -W -f='${Status}' newrelic-php5 2>/dev/null | grep -c "ok installed") -eq 0 ];
+        then
+            echo "1" > "/usr/local/psa/var/modules/new-relic/removepackageapm"
+        else
+            echo "0" > "/usr/local/psa/var/modules/new-relic/removepackageapm"
+        fi
     fi
 fi
